+++ Do-It/View/OrganizationSettingsTableViewCell.swift
@@ -45,80 +45,3 @@ class OrganizationSettingsTableViewCell: UITableViewCell {
 
 }
 
-protocol PrioritySettingSegmentTableViewCellDelegate: AnyObject {
-    func didSelectPriority(_ cell: PrioritySettingSegmentTableViewCell)
-}
-
-class PrioritySettingSegmentTableViewCell: OrganizationSettingsTableViewCell {
-    var doItPriority: DoItPriority?
-    weak var delegate: PrioritySettingSegmentTableViewCellDelegate?
-
-    @IBAction override func selectSegment(sender: UISegmentedControl) {
-        switch sender.selectedSegmentIndex {
-        case 0:
-            doItPriority = nil
-        case 1:
-            doItPriority = .low
-        case 2:
-            doItPriority = .default
-        case 3:
-            doItPriority = .high
-        default:
-            fatalError()
-        }
-
-        delegate?.didSelectPriority(self)
-    }
-
-}
-
-protocol GroupSettingSegmentTableViewCellDelegate: AnyObject {
-    func didSelectGrouping(_ cell: GroupSettingSegmentTableViewCell)
-}
-
-class GroupSettingSegmentTableViewCell: OrganizationSettingsTableViewCell {
-    var doItGroupSetting: GroupingSetting = .dueDate
-    weak var delegate: GroupSettingSegmentTableViewCellDelegate?
-
-    @IBAction override func selectSegment(sender: UISegmentedControl) {
-        switch sender.selectedSegmentIndex {
-        case 0:
-            doItGroupSetting = .dueDate
-        case 1:
-            doItGroupSetting = .course
-        case 2:
-            doItGroupSetting = .priority
-        default:
-            fatalError()
-        }
-
-        delegate?.didSelectGrouping(self)
-    }
-}
-
-protocol SortSettingSegmentTableViewCellDelegate: AnyObject {
-    func didSelectSort(_ cell: SortSettingSegmentTableViewCell)
-}
-
-class SortSettingSegmentTableViewCell: OrganizationSettingsTableViewCell {
-    var doItSortSetting: SortSetting?
-    weak var delegate: SortSettingSegmentTableViewCellDelegate?
-
-    @IBAction override func selectSegment(sender: UISegmentedControl) {
-        switch sender.selectedSegmentIndex {
-        case 0:
-            doItSortSetting = nil
-        case 1:
-            doItSortSetting = .dueDate
-        case 2:
-            doItSortSetting = .course
-        case 3:
-            doItSortSetting = .priority
-        default:
-            fatalError()
-        }
-
-        delegate?.didSelectSort(self)
-    }
-
-}
